+++ clevertap-core/src/main/java/com/clevertap/android/sdk/CleverTapFactory.kt
@@ -508,6 +508,7 @@ internal object CleverTapFactory {
             SYSTEM,
             networkMonitor,
             pipManager,
+            validationResultStack,
         )
         controllerManager.inAppController = inAppController
 
@@ -543,7 +544,7 @@ internal object CleverTapFactory {
         val networkScope = NetworkScope()
 
         val inboxFetchCall = InboxFetchCall(
-            ctApi = ctApiWrapper.ctApi,
+            ctApiProvider = { ctApiWrapper.ctApi },
             queueHeaderBuilder = queueHeaderBuilder,
             coreMetaData = coreMetaData,
             packageName = context.packageName,
@@ -565,7 +566,7 @@ internal object CleverTapFactory {
 
         val inboxDeleteCoordinator = InboxDeleteCoordinator(
             networkScope = networkScope,
-            ctApi = ctApiWrapper.ctApi,
+            ctApiProvider = { ctApiWrapper.ctApi },
             queueHeaderBuilder = queueHeaderBuilder,
             dbAdapterProvider = { databaseManager.loadDBAdapter(context) },
             coreMetaData = coreMetaData,